Add schema validation for json data (#167)
* remove useless item "count" from genome-bgc mappings

* add schemas

* include schemas as package data

* add tests for schemas

* add json validation to GenomeStatus methods

* add json validation for genome bgc mappings

* add json validation for strain mappings

* disallow empty list of bgc ids for genome bgc mappings

* add version to test data to match schema

* add validation for PODP json data

* fix podp test data to pass json validation

* fix installation failure cased by deprecated sklearn issue in bigscape

* update test json data to match schema requirement

* fix data frame dtype bug to pass unit test for `ClassMatches`

+++ src/nplinker/class_info/class_matches.py
@@ -245,7 +245,7 @@ class ClassMatches:
                     class_matching_tables[chem_key] = {}
                     class_matching_counts[chem_key] = {}
                 # add matching tables as DataFrames
-                counts_df = pd.DataFrame.from_dict(counts, dtype=int)
+                counts_df = pd.DataFrame.from_dict(counts)
                 class_matching_tables[bgc_key][chem_key] = (
                     counts_df / counts_df.sum(axis=0)).fillna(0)
                 class_matching_counts[bgc_key][chem_key] = counts_df.fillna(0)
